GET /account/details props from MyAcc to InfoDiv
